Add data models
This commit adds Player, Scout, Report, PositionReport, PitcherReport,
and Pitch data models.

To model reports, I used a newer Rails feature that was introduced in
Rails 6 called `Delegated Types`. Delegated Types provides an
alternative to Single Table Inheritance for modeling inheritance in
Rails. It allows for the reports table to hold the shared data that
position_reports and pitcher_reports have in common and allows the child
models to have their own DB tables with their specific data attributes.
This way there will not be a large shared table with many `null` values
as there would be when using STI.

For more information:

https://api.rubyonrails.org/classes/ActiveRecord/DelegatedType.html